Fix version installs with justhub repo
If we specify a version of the package in the justhub repo, we need to make sure
we specify the same version for both the haskell meta package and the
haskell-min meta package. Otherwise it will still install the latest version of
ghc. Added tests to ensure the correct ghc version gets installed.